GradleTest: rename to GradleLoggingSubmoduleTest
